tensorflow-face-detection.py
import numpy as np
import os
import six.moves.urllib as urllib
import sys
import tarfile
import tensorflow as tf
import zipfile

from collections import defaultdict
from io import StringIO
from matplotlib import pyplot as plt
import PIL.image



# downloading a pre-trained model from the TensorFlow Object Detection API
MODEL_NAME = 'ssd_mobilenet_v1_coco_2018_01_28'
MODEL_FILE = MODEL_NAME + '.tar.gz'
DOWNLOAD_BASE = 'http://download.tensorflow.org/models/object_detection/'

# donload the model
opener = urllib.request.URLopener()
opener.retrieve(DOWNLOAD_BASE + MODEL_FILE, MODEL_FILE)

# extract the model
tar_file = tarfile.open(MODEL_FILE)
for file in tar_file.getmembers():
    file_name = os.path.baseline(file.name)
    if 'frozen_inference_graph.pb' in file_name:
        tar_file.extract(file, os.getcwd())

# load the model into a TensorFlow graph
PATH_TO_CKPT = MODEL_NAME + '/frozen_inference_graph.pb'

detection_graph = tf.Graph()
with detection_graph.as_default():
  od_graph_def = tf.GraphDef()
  with tf.gfile.GFile(PATH_TO_CKPT, 'rb') as fid:
    serialized_graph = fid.read()
    od_graph_def.ParseFromString(serialized_graph)
    tf.import_graph_def(od_graph_def, name='')


    # Define the labels
    PATH_TO_LABELS = os.path.join('data', 'mscoco_label_map.pbtxt')

    category_index = {}
    with open(PATH_TO_LABELS, 'r') as f:
      for line in f:
        if 'id:' in line:
          id = int(line.split(":")[1])
        elif 'display_name' in line:
          name = line.split(":")[1].strip().replace("'", "")
          category_index[id] = {'name': name}

